+++ scripts/tests/harness.sh
@@ -4,7 +4,6 @@
 # executed directly.
 
 export FOREST_CHAIN_INDEXER_ENABLED="1"
-export FOREST_ETH_MAPPINGS_RANGE="300"
 
 export FOREST_PATH="forest"
 export FOREST_CLI_PATH="forest-cli"
@@ -29,7 +28,7 @@ function forest_download_and_import_snapshot {
 
 function forest_download_and_import_snapshot_with_f3 {
   echo "Downloading v1 snapshot"
-  aria2c -x5 https://forest-archive.chainsafe.dev/latest/calibnet/ -o v1.forest.car.zst
+  aria2c -x5 https://forest-archive.chainsafe.dev/latest-v1/calibnet -o v1.forest.car.zst
   echo "Inspecting v1 snapshot"
   $FOREST_TOOL_PATH archive info v1.forest.car.zst
   echo "Downloading F3 snapshot"
@@ -59,14 +58,17 @@ function backfill_db {
   snapshot_epoch=$(get_epoch_from_car_db)
   echo "Snapshot epoch: $snapshot_epoch"
 
-  # Default to 300 if no argument is provided
+  # Return an error if no argument is provided
+  if [[ -z "$1" ]]; then
+    echo "Error: No argument provided. Please provide the backfill epochs."
+    return 1
+  fi
+
+  # Use the provided argument for backfill epochs
   local backfill_epochs
-  backfill_epochs=${1:-300}
+  backfill_epochs=$1
 
-  local to_epoch
-  to_epoch=$((snapshot_epoch - backfill_epochs))
-
-  $FOREST_TOOL_PATH index backfill --chain calibnet --from "$snapshot_epoch" --to "$to_epoch"
+  $FOREST_TOOL_PATH index backfill --chain calibnet --from "$snapshot_epoch" --n-tipsets "$backfill_epochs"
 }
 
 function forest_check_db_stats {