+++ src/eth/homestead_transaction.rs
@@ -108,10 +108,7 @@ impl EthLegacyHomesteadTxArgs {
                 *value = 1
             };
         } else {
-            bail!(
-                "invalid 'v' value: expected 27 or 28, got {}",
-                v_value.to_string()
-            );
+            bail!("invalid 'v' value: expected 27 or 28, got {}", v_value);
         }
 
         Ok(sig)